[refactor/chore] code-cleanup continued; merging with upstream
